third_party/blink/web_tests/external/wpt/client-hints/service-workers/service-worker-passthrough-request.https.html

<!DOCTYPE html>
<title>Simple test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test((t) => {
  navigator.serviceWorker.register('resources/passthrough-request.js')
  .then((e) =>{
    var popup_window = window.open("/client-hints/service-workers/resources/echo-hint-in-html.py");
    assert_not_equals(popup_window, null, "Popup windows not allowed?");
    popup_window.addEventListener('load', (e) => {
      t.step(()=>{assert_equals(popup_window.document.body.textContent, "PASS")});
      t.done();
    });
  })
}, "Client Hint/Service Worker interaction: service worker successfully passes hints through to new fetch");
</script>
