third_party/blink/web_tests/virtual/legacy-client-hints-no-fp-delegation/external/wpt/client-hints/service-workers/service-worker-passthrough-request.https-expected.txt

This is a testharness.js-based test.
FAIL Client Hint/Service Worker interaction: service worker successfully passes hints through to new fetch assert_equals: expected "PASS" but got "FAIL"
Harness: the test ran to completion.
